requirements.txt
pytest
mock
repoze.lru
